Move tests into correct package
